feat(core): sort SWC nodes by default
Enable the `sort_nodes` option for the `read_swc` function by default.
This also affects `Tree.from_swc`, `Tree.from_eswc`, `Population.from_swc`,
and many other related functions.

This option is necessary because some SWC files do not have continuous IDs,
which is a critical assumption in our library. Previously, it was disabled by
default for performance reasons. However, we’ve received many bug reports that
were difficult to track down. To improve usability, we are now enabling this
option by default, hoping it will prevent further issues.

This change should not break any existing scripts, unless they explicitly rely
on the original ID order (which they should NOT). We also hope to improve the
performance of this option in the future.